feat: chart fix

+++ Assets/02. Scripts/RealEstate/RealEstateData.cs
@@ -5,18 +5,15 @@ using UnityEngine;
 public class RealEstateData {
     public string id;
     public string name;
-    public int price;
+
+    public long price; 
+
     public float monthlyYield;
     public string ownedText;
-
-    // 추후 확장용
     public bool owned;
-
-    // 이미지
-    public string imageKey;   // ex) "estate_500"
+    public string imageKey;
     [System.NonSerialized]
-    public Sprite image;      // 런타임 로드용
-
+    public Sprite image;
     public DateTime buyDate;
     public DateTime nextIncomeDate;
     public DateTime lastPriceUpdateDate;
